Caption/requirement.txt
annotated-types==0.7.0
anyio==4.5.2
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
groq==0.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
pillow==10.4.0
pydantic==2.10.1
pydantic_core==2.27.1
sniffio==1.3.1
sympy==1.13.3
torch==2.4.0
torchvision==0.19.0
typing_extensions==4.12.2
Werkzeug==3.0.6
zipp==3.20.2
